code refactoring - use spring to instantiate the resolver
