+++ config/src/main/java/com/quorum/tessera/config/JdbcConfig.java
@@ -1,21 +1,16 @@
 package com.quorum.tessera.config;
 
-import com.quorum.tessera.config.adapters.EncryptedStringAdapter;
-
 import javax.validation.constraints.NotNull;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 public class JdbcConfig extends ConfigItem {
 
     @XmlElement private String username;
 
-    @XmlElement(type = String.class)
-    @XmlJavaTypeAdapter(value = EncryptedStringAdapter.class)
-    private String password;
+    @XmlElement private String password;
 
     @NotNull
     @XmlElement(required = true)
